// mail/models/MailAccounts.ts
/**
 * Liara Mail API Documentaion
 * A fully featured mail delivery platform for incoming & outgoing e-mail
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { MailAccountsData } from '../models/MailAccountsData';
import { HttpFile } from '../http/http';

export class MailAccounts {
    'status'?: string;
    'data'?: MailAccountsData;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "MailAccountsData",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return MailAccounts.attributeTypeMap;
    }

    public constructor() {
    }
}