// mail/models/MailAttachments.ts
/**
 * Liara Mail API Documentaion
 * A fully featured mail delivery platform for incoming & outgoing e-mail
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { MailAttachmentsData } from '../models/MailAttachmentsData';
import { HttpFile } from '../http/http';

export class MailAttachments {
    'status'?: string;
    'data'?: MailAttachmentsData;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "MailAttachmentsData",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return MailAttachments.attributeTypeMap;
    }

    public constructor() {
    }
}